+++ src/App.jsx
@@ -1,4 +1,4 @@
-import { useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import Header from "./components/Header";
 import Main from "./components/Main";
 import Footer from "./components/Footer";
@@ -6,18 +6,17 @@ import Modal from "./components/Modal";
 import "./styles/App.css";
 
 export default function App() {
-  const [newGame, setNewGame] = useState(false);
   const [gameOver, setGameOver] = useState(false);
   const [currentScore, setCurrentScore] = useState(0);
   const [bestScore, setBestScore] = useState(currentScore);
   const gameWinRef = useRef(false);
-  console.log("gameWinRef.current", gameWinRef.current);
+  const gameStatusRef = useRef(null);
   return (
     <div id="app">
       <Header currentScore={currentScore} bestScore={bestScore} />
       <Main
+        gameStatusRef={gameStatusRef}
         gameWinRef={gameWinRef}
-        newGame={newGame}
         gameOver={gameOver}
         currentScore={currentScore}
         bestScore={bestScore}
@@ -26,17 +25,23 @@ export default function App() {
         setBestScore={setBestScore}
       />
       <Footer />
-      {/* {gameOver && (
+      {gameOver && (
         <Modal
           gameWinRef={gameWinRef}
           openModal={gameOver}
-          closeModal={(newGame) => {
+          closeModal={(isNewGame) => {
             gameWinRef.current = false;
+            const { newGame, tryAgain } = gameStatusRef.current;
+            if (isNewGame) {
+              newGame.fn(newGame.args);
+            } else {
+              const { callback, cards } = tryAgain.args;
+              tryAgain.fn(callback(cards));
+            }
             setGameOver(false);
-            setNewGame(newGame);
           }}
         />
-      )} */}
+      )}
     </div>
   );
 }
@@ -55,10 +60,13 @@ export default function App() {
  *    Best score
  * Card container
  *    Card
- * Card count slider(?)
+ * Card count slider (?)
  *    Min/max
  *    Card count
  * A way to fetch data from an API
+ * Game status pop up (?)
+ *    Try Again Button
+ *    New Game Button
  *
  * Behavior:
  *    On load
@@ -71,17 +79,23 @@ export default function App() {
  *      Best score is set to 0
  *      When an image is clicked:
  *          If clicked image has not been clicked, current score increments by one
+ *              If current score is equal to card count
+ *                  Set game over state to true (?)
+ *              Shuffle images (re-render)
  *          If clicked image has already been clicked
+ *              Set game over state to true (?)
  *              If current score is greater than best score
  *                  Set best score to current score
- *               Current score resets to 0
- *          All images will shuffle (re-render)
+ *                  Current score resets to 0
  *
  *    On change card count slider
  *      Card count is set to the value of card count slider
  *      Render cards based on card count's value
  *      Score is set to 0
  *      Best score is set to 0
+ *
+ *    If game over state is true (?)
+ *      Render a pop up (dialog element)
  *
  * Possible problems:
  *    Duplicate images may occur when fetching X amount of 'random' images from an API